chore(cart) : Rest Doc 테스트코드 클래스명 변경
